requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==10.0.0
Babel==2.13.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
beartype==0.17.2
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
dlib==19.24.2
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.18.1
filelock==3.12.4
flatbuffers==24.3.25
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
fvcore==0.1.5.post20221221
gmpy2==2.1.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
h5dict==0.2.3
h5py==3.10.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.5.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.9
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.4.0
jupyter-console==6.6.3
jupyter_core==5.4.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.2.post1
lightning==2.1.3
lightning-utilities==0.10.0
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.18.1
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.17.0
pooch==1.8.2
portalocker==2.8.2
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-fid==0.3.0
pytorch-lightning==2.1.3
pytorch3d==0.7.4
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.7.24
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.1
scikit-video==1.1.11
scipy==1.10.1
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==1.12.1
torch-tb-profiler==0.4.3
torchaudio==0.12.1
torchmetrics==1.2.1
torchvision==0.13.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.44.0
triton==2.1.0
tslearn==0.6.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
typing-utils==0.1.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
wheel==0.41.2
widgetsnbextension==4.0.9
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0
